@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/hopr-core-ethereum",
-  "version": "1.68.0-next.3",
+  "version": "1.69.0-next.11",
   "description": "",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
@@ -15,21 +15,21 @@
     "build": "yarn clean && yarn build:types && mkdir -p ./lib/tsc/web3 && cp -R ./src/tsc/web3 ./lib/tsc/web3 && tsc",
     "prepublishOnly": "yarn build"
   },
-  "files": [
-    "lib/",
-    "!*.spec.ts",
-    "!*.spec.d.ts",
-    "!*.spec.js",
-    "!*.spec.js.map"
-  ],
   "engines": {
-    "node": "12",
+    "node": "14",
     "yarn": "1"
   },
+  "files": [
+    "lib",
+    "!**/*.spec.ts",
+    "!**/*.spec.d.ts",
+    "!**/*.spec.js",
+    "!**/*.spec.js.map"
+  ],
   "dependencies": {
-    "@hoprnet/hopr-ethereum": "1.68.0-next.3",
-    "@hoprnet/hopr-utils": "1.68.0-next.3",
-    "async-mutex": "^0.2.6",
+    "@hoprnet/hopr-ethereum": "1.69.0-next.11",
+    "@hoprnet/hopr-utils": "1.69.0-next.11",
+    "async-mutex": "^0.3.0",
     "bn.js": "^5.1.2",
     "chalk": "^4.1.0",
     "debug": "^4.3.1",
@@ -45,15 +45,15 @@
     "web3-eth-contract": "^1.3.3"
   },
   "devDependencies": {
-    "@hoprnet/hopr-core-connector-interface": "1.68.0-next.3",
-    "@hoprnet/hopr-demo-seeds": "1.68.0-next.3",
-    "@hoprnet/hopr-testing": "1.68.0-next.3",
+    "@hoprnet/hopr-core-connector-interface": "1.69.0-next.11",
+    "@hoprnet/hopr-demo-seeds": "1.69.0-next.11",
+    "@hoprnet/hopr-testing": "1.69.0-next.11",
     "@types/debug": "^4.1.5",
     "@types/keccak": "^3.0.1",
     "@types/levelup": "^4.3.0",
     "@types/memdown": "^3.0.0",
     "@types/mocha": "^8.2.0",
-    "@types/node": "^12",
+    "@types/node": "14",
     "@types/secp256k1": "^4.0.1",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
