@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/hopr-core",
-  "version": "1.68.0-next.3",
+  "version": "1.69.0-next.11",
   "description": "Privacy-preserving messaging protocol with incentivations for relay operators",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
@@ -12,15 +12,8 @@
   ],
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
-  "files": [
-    "lib/",
-    "!*.spec.ts",
-    "!*.spec.d.ts",
-    "!*.spec.js",
-    "!*.spec.js.map"
-  ],
   "engines": {
-    "node": "12",
+    "node": "14",
     "yarn": "1"
   },
   "scripts": {
@@ -30,10 +23,17 @@
     "testOnce": "node --trace-warnings --unhandled-rejections=strict node_modules/.bin/mocha --reporter=tap --full-trace --exit",
     "prepublishOnly": "yarn build"
   },
+  "files": [
+    "lib",
+    "!**/*.spec.ts",
+    "!**/*.spec.d.ts",
+    "!**/*.spec.js",
+    "!**/*.spec.js.map"
+  ],
   "dependencies": {
-    "@hoprnet/hopr-connect": "0.2.9",
-    "@hoprnet/hopr-core-ethereum": "1.68.0-next.3",
-    "@hoprnet/hopr-utils": "1.68.0-next.3",
+    "@hoprnet/hopr-connect": "0.2.12",
+    "@hoprnet/hopr-core-ethereum": "1.69.0-next.11",
+    "@hoprnet/hopr-utils": "1.69.0-next.11",
     "abort-controller": "^3.0.0",
     "bn.js": "5.1.2",
     "chalk": "~4.1.0",
@@ -43,11 +43,11 @@
     "it-pipe": "^1.1.0",
     "leveldown": "~5.6.0",
     "levelup": "~4.4.0",
-    "libp2p": "0.29",
-    "libp2p-crypto": "0.18.0",
+    "libp2p": "0.30",
+    "libp2p-crypto": "0.19.0",
     "libp2p-kad-dht": "0.20.6",
-    "libp2p-mplex": "0.10.1",
-    "libp2p-secio": "0.13.1",
+    "libp2p-mplex": "0.10.2",
+    "libp2p-noise": "^2.0.4",
     "multiaddr": "8.1.2",
     "multihashes": "3.1.2",
     "peer-id": "0.14.3",
@@ -56,10 +56,10 @@
     "secp256k1": "~4.0.2"
   },
   "devDependencies": {
-    "@hoprnet/hopr-core-connector-interface": "1.68.0-next.3",
-    "@hoprnet/hopr-demo-seeds": "1.68.0-next.3",
-    "@hoprnet/hopr-ethereum": "1.68.0-next.3",
-    "@hoprnet/hopr-testing": "1.68.0-next.3",
+    "@hoprnet/hopr-core-connector-interface": "1.69.0-next.11",
+    "@hoprnet/hopr-demo-seeds": "1.69.0-next.11",
+    "@hoprnet/hopr-ethereum": "1.69.0-next.11",
+    "@hoprnet/hopr-testing": "1.69.0-next.11",
     "@types/chai-as-promised": "^7.1.3",
     "@types/debug": "^4.1.5",
     "@types/err-code": "^2.0.0",
@@ -67,15 +67,15 @@
     "@types/levelup": "^4.3.0",
     "@types/memdown": "^3.0.0",
     "@types/mocha": "^8.2.0",
-    "@types/node": "^12",
+    "@types/node": "14",
     "@types/read": "^0.0.28",
     "@types/secp256k1": "~4.0.1",
     "@types/sinon": "^9.0.8",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "libp2p-tcp": "0.15.2",
+    "libp2p-tcp": "0.15",
     "memdown": "^5.1.0",
-    "mocha": "^8.2.0",
+    "mocha": "^8.2.1",
     "rimraf": "^3.0.2",
     "sinon": "^9.0.8",
     "typescript": "^4.1"
