@@ -1,7 +1,7 @@
 {
   "name": "@hoprnet/hopr-utils",
   "description": "HOPR-based utilities to process multiple data structures",
-  "version": "1.68.0-next.3",
+  "version": "1.69.0-next.11",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
   "license": "LGPL-3.0-only",
@@ -14,17 +14,18 @@
     "build": "yarn clean && tsc -p . --noEmit false",
     "prepublishOnly": "yarn build"
   },
-  "files": [
-    "lib/",
-    "!lib/**/*.spec.*",
-    "README.md",
-    "package.json",
-    "yarn.lock"
-  ],
   "engines": {
-    "node": ">=12.9.1"
+    "node": "14"
   },
+  "files": [
+    "lib",
+    "!*.spec.ts",
+    "!*.spec.d.ts",
+    "!*.spec.js",
+    "!*.spec.js.map"
+  ],
   "dependencies": {
+    "abort-controller": "~3.0.0",
     "bignumber.js": "^9.0.0",
     "libp2p-crypto": "0.18.0",
     "multiaddr": "8.1.2",
@@ -34,8 +35,9 @@
   },
   "devDependencies": {
     "@types/mocha": "^8.2.0",
-    "@types/node": "^12",
+    "@types/node": "14",
     "mocha": "^8.2.0",
+    "rewiremock": "^3.14.3",
     "rimraf": "^3.0.2",
     "ts-node": "^9.0.0",
     "typescript": "^4.1"
