@@ -1,12 +1,9 @@
 import LibP2P from 'libp2p'
 import Multiaddr from 'multiaddr'
 import PeerId from 'peer-id'
-// @ts-ignore
-import TCP = require('libp2p-tcp')
-// @ts-ignore
-import MPLEX = require('libp2p-mplex')
-// @ts-ignore
-import SECIO = require('libp2p-secio')
+const TCP = require('libp2p-tcp')
+const MPLEX = require('libp2p-mplex')
+import { NOISE } from 'libp2p-noise'
 
 /**
  * Informs each node about the others existence.
@@ -37,11 +34,11 @@ export type LibP2PMocks = {
 export async function generateLibP2PMock(addr = '/ip4/0.0.0.0/tcp/0'): Promise<LibP2PMocks> {
   const node = await LibP2P.create({
     peerId: await PeerId.create({ keyType: 'secp256k1' }),
-    addresses: { listen: [Multiaddr(addr)] },
+    addresses: { listen: [addr] },
     modules: {
       transport: [TCP],
       streamMuxer: [MPLEX],
-      connEncryption: [SECIO]
+      connEncryption: [NOISE]
     }
   })
 
