@@ -1,7 +1,6 @@
 import PeerId from 'peer-id'
 import { keys as libp2p_crypto } from 'libp2p-crypto'
-// @ts-ignore
-import Multihash = require('multihashes')
+import Multihash from 'multihashes'
 
 /**
  * Takes a peerId and returns a peerId with the public key set to the corresponding
