@@ -1,8 +1,8 @@
-import type { Handler } from 'libp2p'
+import type { Connection, MuxedStream } from 'libp2p'
 
 interface AbstractInteraction {
   protocols: string[]
-  handler(struct: Handler): void
+  handler(struct: { connection: Connection; stream: MuxedStream; protocol: string }): void
   interact(...props: any[]): any
 }
 
