@@ -3,18 +3,18 @@ import type Hopr from '..'
 import PeerId from 'peer-id'
 
 import { PaymentInteractions } from './payments'
-import { NetworkInteractions } from './network'
 import { PacketInteractions } from './packet'
+import { Heartbeat } from './network/heartbeat'
 import { Mixer } from '../mixer'
 
 class Interactions<Chain extends HoprCoreConnector> {
   public payments: PaymentInteractions<Chain>
-  public network: NetworkInteractions
+  heartbeat: Heartbeat
   public packet: PacketInteractions<Chain>
 
   constructor(node: Hopr<Chain>, mixer: Mixer<Chain>, heartbeat: (remotePeer: PeerId) => void) {
     this.payments = new PaymentInteractions(node)
-    this.network = new NetworkInteractions(node._libp2p, heartbeat)
+    this.heartbeat = new Heartbeat(node._libp2p, heartbeat)
     this.packet = new PacketInteractions(node, mixer)
   }
 }
