@@ -10,7 +10,7 @@ import { UnacknowledgedTicket } from '.'
 import { NODE_SEEDS } from '@hoprnet/hopr-demo-seeds'
 
 import { Types, Utils } from '@hoprnet/hopr-core-ethereum'
-import { privKeyToPeerId } from '../../utils'
+import { privKeyToPeerId } from '@hoprnet/hopr-utils'
 import { toU8a, u8aConcat, u8aToNumber } from '@hoprnet/hopr-utils'
 
 import LevelUp from 'levelup'
