@@ -4,7 +4,7 @@ import assert from 'assert'
 import { HEARTBEAT_INTERVAL, NETWORK_QUALITY_THRESHOLD } from '../constants'
 import sinon from 'sinon'
 import { fakePeerId } from '../test-utils'
-import type PeerId from 'peer-id'
+import PeerId from 'peer-id'
 
 describe('unit test heartbeat', async () => {
   let heartbeat: Heartbeat
@@ -16,9 +16,9 @@ describe('unit test heartbeat', async () => {
     interact: sinon.fake.resolves(true)
   } as any
 
-  beforeEach(() => {
+  beforeEach(async () => {
     clock = sinon.useFakeTimers(Date.now())
-    peers = new NetworkPeerStore([])
+    peers = new NetworkPeerStore([], [await PeerId.create({ keyType: 'secp256k1' })])
     heartbeat = new Heartbeat(peers, interaction, hangUp)
   })
 
@@ -50,10 +50,11 @@ describe('unit test heartbeat', async () => {
   it('test heartbeat flow', async () => {
     let generateMock = (i: string | number) => {
       let id = fakePeerId(i)
-      let peers = new NetworkPeerStore([])
+      let peers = new NetworkPeerStore([], [id])
       let heartbeat = new Heartbeat(peers, interaction, hangUp)
       return { peers, interaction, id, heartbeat }
     }
+
     let alice = generateMock(1)
     let bob = generateMock(2)
     let chris = generateMock(3)
@@ -87,9 +88,9 @@ describe('unit test heartbeat', async () => {
     // Chris dies, alice heartbeats again
     alice.interaction.interact = sinon.fake((id: PeerId) => {
       if (id.equals(chris.id)) {
-        throw new Error('FAIL')
+        return Promise.resolve(-1)
       }
-      return Promise.resolve()
+      return Promise.resolve(0)
     })
 
     clock.tick(HEARTBEAT_INTERVAL * 2)
